pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "layr-sdk"
version = "0.1.0"
description = "Observability for AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = [
    "ai", "agents", "observability",
    "monitoring", "llm", "langchain",
    "opentelemetry", "tracing"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "httpx>=0.24.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.optional-dependencies]
langchain = ["langchain>=0.1.0"]
crewai = ["crewai>=0.1.0"]
autogen = ["pyautogen>=0.2.0"]
datadog = ["datadog>=0.47.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://getlayr.io"
Documentation = "https://docs.getlayr.io"
Repository = "https://github.com/getlayr/layr-sdk"
Issues = "https://github.com/getlayr/layr-sdk/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["layr*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=layr --cov-report=term-missing"
